FamilyDBLogic

+++ MyApp/app/src/main/java/com/example/janari/SimpleDailyBudgetApp/CalculateDailySumClass.java
@@ -4,7 +4,7 @@ package com.example.janari.SimpleDailyBudgetApp;
 public class CalculateDailySumClass {
 
 
-    public static double calculateSum (double value, double sum, double days){
+    public static double calculateSum (double value, double sum, int days){
 
         return (value - sum)/ days;
     }
